news-summary/claude/requirements_claude.txt
flask
anthropic
google-cloud-firestore
gunicorn